Move OptionValueAccessor to options package
